XMLHttpRequest/open-send-open.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() - send() - open()</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest(),
            result = [],
            expected = [1,1,1,1]
        client.onreadystatechange = function() {
          test.step(function() {
            result.push(client.readyState)
          })
        }
        client.open("GET", "folder.txt")
        client.send()
        client.open("GET", "folder.txt")
        assert_array_equals(result, expected)
        test.done()
      })
    </script>
  </body>
</html>
